backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
# --- New dependencies for Whisper/Audio processing ---
# Install the whisper package
openai-whisper
# Install dependencies for pydub to handle formats (requires ffmpeg system dependency)
pydub
pytest>=8.2.0
pytest-asyncio>=0.23.6
httpx>=0.27.0
pytest-cov>=5.0.0
coverage>=7.5.1
# Linters/Tools (for pre-commit/CI)
black>=24.4.2
isort>=5.13.2
flake8>=7.1.0
mypy>=1.10.0
pydantic>=2.0.0